// src/app/models/location.ts
import { BaseModel } from './base-model';

export class Location extends BaseModel {
    id: string;
    companyName: string;
    locationName: string;
    dateEntered: Date;
    dateModified: Date;
    modifiedUserId: string;
    createdBy: string;
    description: string;
    salesRepId: string;
    salesRep: string;
    locationType: string;
    industry: string;
    annualRevenue: string;
    phoneFax: string;
    billingAddressStreet: string;
    billingAddressStreet2: string;
    billingAddressCity: string;
    billingAddressState: string;
    billingProvice: string;
    billingAddressPostalcode: string;
    billingAddressCountry: string;
    rating: string;
    phoneOffice: string;
    phoneAlternate: string;
    website: string;
    ownership: string;
    employees: string;
    tickerSymbol: string;
    shippingAddressStreet: string;
    shippingAddressStreet2: string;
    shippingAddressCity: string;
    shippingAddressState: string;
    shippingProvince: string;
    shippingAddressPostalcode: string;
    shippingAddressCountry: string;
    deliveryContact: string;
    qboSalesTax: string;
    qbdSalesTax: string;

    setData(data) {
        this.setString(data, 'id');
        this.setString(data, 'companyName');
        this.setString(data, 'locationName');
        this.setDate(data, 'dateEntered');
        this.setDate(data, 'dateModified');
        this.setString(data, 'modifiedUserId');
        this.setString(data, 'createdBy');
        this.setString(data, 'description');
        this.setString(data, 'salesRepId');
        this.setString(data, 'salesRep');
        this.setString(data, 'locationType');
        this.setString(data, 'industry');
        this.setString(data, 'annualRevenue');
        this.setString(data, 'phoneFax');
        this.setString(data, 'billingAddressStreet');
        this.setString(data, 'billingAddressStreet2');
        this.setString(data, 'billingAddressCity');
        this.setString(data, 'billingAddressState');
        this.setString(data, 'billingProvice');
        this.setString(data, 'billingAddressPostalcode');
        this.setString(data, 'billingAddressCountry');
        this.setString(data, 'rating');
        this.setString(data, 'phoneOffice');
        this.setString(data, 'phoneAlternate');
        this.setString(data, 'website');
        this.setString(data, 'ownership');
        this.setString(data, 'employees');
        this.setString(data, 'tickerSymbol');
        this.setString(data, 'shippingAddressStreet');
        this.setString(data, 'shippingAddressStreet2');
        this.setString(data, 'shippingAddressCity');
        this.setString(data, 'shippingAddressState');
        this.setString(data, 'shippingProvince');
        this.setString(data, 'shippingAddressPostalcode');
        this.setString(data, 'shippingAddressCountry');
        this.setString(data, 'deliveryContact');
        this.setString(data, 'qboSalesTax');
        this.setString(data, 'qbdSalesTax');
    }
}